prov/efa: rename and move efa_alloc_info()
efa_alloc_info() allocates fi_info objects that is linked in
efa_util_prov.info. These info objects are used to verify
the info passed from user, and are generally refered to as
"prov_info" objects in libfabric utility code.

To make the code more clear, this patch rename the function
to efa_prov_info_alloc() and moved it and related functions
to a seperate file efa_prov_info.c. It also introduced
a header file efa_prov_info.h that has the definition.
